requirements.txt
aiml; python_version <= '2.7'
archey4; python_version >= '3.5' and sys_platform != 'darwin'
beautifulsoup4
colorama
distro
basketball_reference_web_scraper
flake8
forex-python
funcsigs; python_version <= '2.7'
googletrans
IMDbPY
instantmusic
mock
nltk
pick
pluginmanager
pycricbuzz
pydoc-markdown
pync==1.6.1
python-dateutil
pytimeparse
pyttsx3
requests[security]
six
sympy
termdown
wikipedia
https://github.com/appi147/pyaiml3/archive/master.zip ; python_version >= '3.5'
speedtest-cli
